Split StateImportResolver to a file
